Change expectation of parsing invalid literals
Currently, the scanner will only collect the valid digits of number
literals, meaning an input like "0xtest" will return an 'empty' error
because the parser thinks there are no digits and that "test" is an
Ident token.
Example of current behavior:

	$ numbrs
	> 0xtest
	Error: failed to parse integer: cannot parse integer from empty string

Instead, all alphanumeric digits should be collected, then an 'invalid
digit' error should be returned if the digits present are invalid.
Example of correct behavior:

    $ numbrs
    > 0xtest
    Error: failed to parse integer: invalid digit found in string

This commit only implements tests that expect the 'correct' behavior. It
doesn't fix the problem.