Add tests for ideal literal scanning
Currently, number literals with invalid characters are scanned anyways
and it's the parser's job to raise an error when parsing them. Instead,
the scanner should detect invalid number literal syntax.

This commit adds tests to the 'scanner' module that test for the ideal
behavior. It also deletes the 'invalid_number_literal' test from the
'parser' module since it's not the parser's job to detect those errors.