Print license notices on command-line
Prints a startup message when numbrs is run and a longer license
notice when the `license` command is issued.

The messages are stored in the `src/{startup,license}_message.txt.in`
files. Some placeholders are available and are filled in by the
build script `build.rs` during compilation.

The list of copyright years is stored as a string in Cargo.toml under
the `package.metadata.copyright-years` key.