Implement unit prefixes and suffixes
Prefixes: Units often have prefixes, such as "kilo" in "kilogram"
meaning thousand. This is now supported. Any identifier that does not
exactly match a variable in the environment will be tested for each
prefix. If a variable with the name matching the identifier minus a
prefix is found, it is scaled according to that prefix's value and
returned.

Suffixes: Units can also have suffixes, typically to make plural
quantities more grammatically correct. For example, "3 inches" means the
"inch" unit; it just has a plural suffix. Units are also matched by
removing any possible suffix and replacing it with the singular suffix
where applicable (e.g. "ies" becomes "y" for "henries").