Add parser/lexer tester examples
This commit adds two "example" applications under the `/examples`
directory. These are more like tests, but they're not automatic
integration tests, so they don't belong in `/tests`.

These two "example" applications will help with debugging, as it is
often useful to see how a given input line was processed when it
provides an unexpected result.

The `/examples/tokenize.rs` application provides the same style of
interface as the main Numbrs executable, except it only tokenizes
(lexes) the input line. It does not parse and evaluate the line. It
prints the list of tokens as the output rather than the evaluated
result.

The `/examples/parse.rs` application is similar, but it parses the line
as well. It prints the AST for the input line as the output. This
application also does not evaluate the input line. It only lexes and
parses it.