Tweak operator binding powers
This commit makes some changes to the binding powers of operators. These
changes don't affect the behavior of Numbrs, but they make the code make
more sense.

Each operator has a base binding power which is a multiple of 10. This
base binding power is given to the associative side and the other side
is given a binding power of the base + 5. For example, the addition
operator is left associative, so it has a left, right binding power of
30, 35.

There is one exception, which is the assignment operator. This is
explained in the README.