Reduce size of EvalError and RuntimeError
When functions return a Result that contains a large error type, enough
memory must be allocated to fit the error type even if it is
infrequently used.

EvalError and RuntimeError both had variants which contained large
structs like Value or Units. This commit wraps those in a Box so they
get stored on the heap. The error structs now only have to store a
pointer to the boxed data.

This change shrinks EvalError and RuntimeError from 184 bytes to 40
bytes and 48 bytes, respectively.