Make unwrap() illegal
This commit enables the clippy::unwrap_used diagnostic and makes it
fatal. This means calling the `unwrap()` function on a `Result` or an
`Option` is no longer allowed.

In cases where it is safe to unwrap, the `allow(clippy::unwrap_used)`
attribute must be specified and a comment must be added to explain why
it is safe to unwrap.

It is also allowable to allow unwrapping for unit tests, as they are
designed to fail on errors, not keep running.