Change package name to lowercase
Most crate or package names are lowercase, so we should follow that
standard. This commit changes the package name from "Numbrs" to
"numbrs".

The program will still be referred to as Numbrs with a capital 'N'. It
is only the Cargo package name that changes.

This commit also changes the build script to capitalize the package name
before using it in substitutions.