Don't make compiler warnings fatal in development
This commit removes the line in `src/lib.rs` which treats compiler
warnings like errors, making them fatal.

It is unhelpful during development for the build to fail on warnings.
It's better to allow the build to continue with warnings while
developing. Failing on warnings should only happen for stable code.

Because of this, the CI pipeline will still fail on warnings, and that
pipeline must pass in order to merge into `master`. So warnings are
still fatal in production, but they will not hinder local development.