Wrap prefix and suffix lists in accessor functions
The list of standard prefixes and standard suffixes are both static
objects. The prefix list must be initialized at runtime, so it must be
wrapped in a lazy initializer. To avoid making the public library API
expose the implementation of this, we wrap both lists in accessor
functions which return a static slice.

This commit changes the public API and therefore requires a minor
version bump.