Isolate Operation into its own sub-module
We put Operation in its own module because the strum macros cause
warnings which cannot be disabled for just the Operation type since they
occur within procedural macros. We must allow the unreachable_patterns
diagnostic for the containing module, so we isolate Operation into its
own module to avoid affecting other code with that attribute.