Don't prioritize implicit operations
Giving implicit operations higher binding power led to inconsistent
syntax problems.

For example,
    1 m / 2 s
should probably be interpreted as
    (1 m) / (2 s)
.

However,
    1/360 circle
should be interpreted as
    (1/360) circle
but with prioritized implicit operations it would be
    1 / (360 circle)
.

This leads to context-specific binding power for implicit operations, so
it's better to not change the binding power of implicit operations at
all.